requirements.txt
spotipy
snowflake-connector-python
python-dotenv
pandas
numpy
scikit-learn
boto3